+++ src/features/XIT/index.ts
@@ -39,6 +39,7 @@ import './PLAN/JH';
 import './PROD/PROD';
 import './PWARN/PWARN';
 import './REP/REP';
+import './REPP/REPP';
 import './SET/SET';
 import './START';
 import './TODO/TODO';