+++ src/store/user-data.ts
@@ -31,6 +31,7 @@ export const initialUserData = deepFreeze({
       threshold: 60,
       offset: 10,
     },
+    repairPlan: {},
     sidebar: [
       ['基地', 'BS'],
       ['总部', 'XIT HQUC'],
@@ -48,6 +49,7 @@ export const initialUserData = deepFreeze({
       ['脚本', 'XIT ACT'],
       ['消耗', 'XIT BURN'],
       ['维护', 'XIT REP'],
+      ['维修预测', 'XIT REPP'],
       ['设置', 'XIT SET'],
       ['帮助', 'XIT HELP'],
       ['计划', 'XIT JH'],